Updated 2 files
Updated event_logger.py to catch an exception that was missed previously.
When trying to delete an ephemeral message, the bot can't find the message, and errors.

Updated suggestions to fix a bug when sending suggestion to internal design, the queue would stop on trying to respond to the interaction a second time.