Updated 3 files
Updated event logger:
Open file if message sent in PG channels.
updated embeds a bit more for error handling.
Created helper function that will add st, nd, rd, or th to the end of numbers, depending on what number it is.

Updated Fun:
added status command. This will be used eventually to show players stats of their character, or the character of others. As well as their town stats, and other useful information in the server.

Updated suggestions:
Created a helper function that will check if we are being ratelimited while updating the views of the suggestion channels. This will enable the staff member to change the topic, while at the same time, understanding what would be wrong if the name didn't change. After the rate limit has passed, the channel should change names.